native: split runtime player subdomain folders
