+++ src/native/engine/engine.c
@@ -15,19 +15,19 @@
 #include "runtime/camera_follow.h"
 #include "runtime/camera_basis.h"
 #include "runtime/move_target_service.h"
-#include "runtime/player_runtime_service.h"
-#include "runtime/player_gateway_abi.h"
+#include "runtime/player/player_runtime_service.h"
+#include "runtime/player/player_gateway_abi.h"
 #include "runtime/render_material.h"
-#include "runtime/player_motion.h"
-#include "runtime/player_motion_host.h"
-#include "runtime/player_api.h"
-#include "runtime/player_build_abi.h"
+#include "runtime/player/player_motion.h"
+#include "runtime/player/player_motion_host.h"
+#include "runtime/player/player_api.h"
+#include "runtime/player/player_build_abi.h"
 #include "runtime/merchant/merchant_query_abi.h"
 #include "runtime/merchant/merchant_trade_domain.h"
-#include "runtime/player_resource_abi.h"
-#include "runtime/player_merchant_adapter.h"
-#include "runtime/player_sync_abi.h"
-#include "runtime/player_registry.h"
+#include "runtime/player/player_resource_abi.h"
+#include "runtime/player/player_merchant_adapter.h"
+#include "runtime/player/player_sync_abi.h"
+#include "runtime/player/player_registry.h"
 #include "runtime/physics_abi.h"
 #include "runtime/input_abi.h"
 #include "runtime/resource_domain.h"
