+++ src/native/engine/runtime/application_service_ports.h
@@ -3,7 +3,7 @@
 
 #include "engine_state.h"
 #include "merchant/merchant_service.h"
-#include "player_motion.h"
+#include "player/player_motion.h"
 
 #ifdef __cplusplus
 extern "C"
