+++ src/native/engine/runtime/engine_composition.h
@@ -2,7 +2,7 @@
 #define BANANA_ENGINE_RUNTIME_ENGINE_COMPOSITION_H
 
 #include "engine_state.h"
-#include "player_motion.h"
+#include "player/player_motion.h"
 
 #ifdef __cplusplus
 extern "C"
