+++ src/native/engine/runtime/player/player_gateway_abi.c
