+++ src/native/engine/runtime/player/player_merchant_adapter.h
@@ -1,7 +1,7 @@
 #ifndef BANANA_ENGINE_RUNTIME_PLAYER_MERCHANT_ADAPTER_H
 #define BANANA_ENGINE_RUNTIME_PLAYER_MERCHANT_ADAPTER_H
 
-#include "application_service_ports.h"
+#include "../application_service_ports.h"
 
 #ifdef __cplusplus
 extern "C"
