+++ src/native/engine/runtime/player/player_motion.h
@@ -1,7 +1,7 @@
 #ifndef BANANA_ENGINE_RUNTIME_PLAYER_MOTION_H
 #define BANANA_ENGINE_RUNTIME_PLAYER_MOTION_H
 
-#include "../world/world.h"
+#include "../../world/world.h"
 #include "player_registry.h"
 
 #ifdef __cplusplus
