+++ src/native/engine/runtime/player/player_resource_abi.c
