+++ src/native/engine/runtime/player/player_resources.c
