+++ src/native/engine/runtime/player/player_resources.h
