+++ tests/native/player_builds_test.c
@@ -1,4 +1,4 @@
-#include "player_builds.h"
+#include "runtime/player/player_builds.h"
 
 #include <stdio.h>
 #include <string.h>
