+++ tests/native/player_sync_staleness_test.c
@@ -1,4 +1,4 @@
-#include "runtime/player_registry.h"
+#include "runtime/player/player_registry.h"
 
 #include <stdio.h>
 #include <stdlib.h>
