+++ tests/native/runtime/player/runtime_player_api_test.c
@@ -1,4 +1,4 @@
-#include "runtime/player_api.h"
+#include "runtime/player/player_api.h"
 
 #include <stdio.h>
 #include <string.h>
