+++ tests/native/runtime/player/runtime_player_motion_host_test.c
@@ -1,4 +1,4 @@
-#include "runtime/player_motion_host.h"
+#include "runtime/player/player_motion_host.h"
 
 #include <math.h>
 #include <stdio.h>
