+++ tests/native/runtime_compile_guard_abi_test.c
@@ -1,12 +1,12 @@
 #include "runtime/controller_kind_domain.h"
 #include "runtime/input_abi.h"
 #include "runtime/move_target_service.h"
-#include "runtime/player_gateway_abi.h"
-#include "runtime/player_build_abi.h"
-#include "runtime/player_builds.h"
+#include "runtime/player/player_gateway_abi.h"
+#include "runtime/player/player_build_abi.h"
+#include "runtime/player/player_builds.h"
 #include "runtime/merchant/merchant_query_abi.h"
 #include "runtime/merchant/merchant_trade_abi.h"
-#include "runtime/player_sync_abi.h"
+#include "runtime/player/player_sync_abi.h"
 #include "runtime/resource_domain.h"
 #include "runtime/terrain_abi.h"
